+++ contracts/contracts/Dependencies/HomoraMath.sol
@@ -2,29 +2,29 @@
 
 pragma solidity 0.6.11;
 
-import './SafeMath.sol';
+import "./SafeMath.sol";
 
 library HomoraMath {
-  using SafeMath for uint;
+  using SafeMath for uint256;
 
-  function divCeil(uint lhs, uint rhs) internal pure returns (uint) {
+  function divCeil(uint256 lhs, uint256 rhs) internal pure returns (uint256) {
     return lhs.add(rhs).sub(1) / rhs;
   }
 
-  function fmul(uint lhs, uint rhs) internal pure returns (uint) {
+  function fmul(uint256 lhs, uint256 rhs) internal pure returns (uint256) {
     return lhs.mul(rhs) / (2**112);
   }
 
-  function fdiv(uint lhs, uint rhs) internal pure returns (uint) {
+  function fdiv(uint256 lhs, uint256 rhs) internal pure returns (uint256) {
     return lhs.mul(2**112) / rhs;
   }
 
   // implementation from https://github.com/Uniswap/uniswap-lib/commit/99f3f28770640ba1bb1ff460ac7c5292fb8291a0
   // original implementation: https://github.com/abdk-consulting/abdk-libraries-solidity/blob/master/ABDKMath64x64.sol#L687
-  function sqrt(uint x) internal pure returns (uint) {
+  function sqrt(uint256 x) internal pure returns (uint256) {
     if (x == 0) return 0;
-    uint xx = x;
-    uint r = 1;
+    uint256 xx = x;
+    uint256 r = 1;
 
     if (xx >= 0x100000000000000000000000000000000) {
       xx >>= 128;
@@ -62,7 +62,7 @@ library HomoraMath {
     r = (r + x / r) >> 1;
     r = (r + x / r) >> 1;
     r = (r + x / r) >> 1; // Seven iterations should be enough
-    uint r1 = x / r;
+    uint256 r1 = x / r;
     return (r < r1 ? r : r1);
   }
 }
