+++ contracts/contracts/Interfaces/ILiquityBase.sol
@@ -4,8 +4,9 @@ pragma solidity 0.6.11;
 
 import "./IPriceFeed.sol";
 
-
 interface ILiquityBase {
-
-    function getEntireSystemDebt() external view returns (uint entireSystemDebt);
+  function getEntireSystemDebt()
+    external
+    view
+    returns (uint256 entireSystemDebt);
 }
