+++ contracts/contracts/Interfaces/ILockupContractFactory.sol
@@ -1,19 +1,24 @@
 // SPDX-License-Identifier: UNLICENSED
 
 pragma solidity 0.6.11;
-    
+
 interface ILockupContractFactory {
-    
-    // --- Events ---
+  // --- Events ---
 
-    event YETITokenAddressSet(address _yetiTokenAddress);
-    event LockupContractDeployedThroughFactory(address _lockupContractAddress, address _beneficiary, uint _unlockTime, address _deployer);
+  event YETITokenAddressSet(address _yetiTokenAddress);
+  event LockupContractDeployedThroughFactory(
+    address _lockupContractAddress,
+    address _beneficiary,
+    uint256 _unlockTime,
+    address _deployer
+  );
 
-    // --- Functions ---
+  // --- Functions ---
 
-    function setYETITokenAddress(address _yetiTokenAddress) external;
+  function setYETITokenAddress(address _yetiTokenAddress) external;
 
-    function deployLockupContract(address _beneficiary, uint _unlockTime) external;
+  function deployLockupContract(address _beneficiary, uint256 _unlockTime)
+    external;
 
-    function isRegisteredLockup(address _addr) external view returns (bool);
+  function isRegisteredLockup(address _addr) external view returns (bool);
 }
