+++ contracts/contracts/TestContracts/ActivePoolTester.sol
@@ -5,13 +5,12 @@ pragma solidity 0.6.11;
 import "../ActivePool.sol";
 
 contract ActivePoolTester is ActivePool {
-    
-    function unprotectedIncreaseYUSDDebt(uint _amount) external {
-        YUSDDebt  = YUSDDebt.add(_amount);
-    }
-    
-    function unprotectedPayable() external payable {
-        // @KingYet: Commented
-        // ETH = ETH.add(msg.value);
-    }
+  function unprotectedIncreaseYUSDDebt(uint256 _amount) external {
+    YUSDDebt = YUSDDebt.add(_amount);
+  }
+
+  function unprotectedPayable() external payable {
+    // @KingYet: Commented
+    // ETH = ETH.add(msg.value);
+  }
 }
